server/ML_part/recommendation_part/app/requirements.txt
fastapi
uvicorn
psycopg2
pandas
python-dotenv
requests
